src/ai/requirements.txt
flask>=2.0.1
werkzeug>=2.0.1
gunicorn>=20.1.0
requests>=2.31.0
google-genai>=0.1.0
langchain>=0.0.1
langchain-core>=0.0.1
langchain-google-genai>=2.0.0
langchain-community>=0.0.1
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
bs4>=0.0.1
beautifulsoup4>=4.10.0
PyPDF2>=3.0.0
supabase>=2.0.0 